Add files via upload
Using Recursion CTE For IT Organizational Hierarchy

Who is your manager?
 I'm going to use the dataset of an IT-organization which is provided in the table employee. The table has the fields ID (ID number of the employee), Name (the employee's name), and Supervisor (ID number of the supervisor).
The IT-organization consists of different roles and levels.


The organization has one IT director (ID=1, Heinz Griesser, Supervisor=0) with many subordinate employees. Under the IT director you can find the IT architecture manager (ID=10, Andreas Sternig, Supervisor=1) with three subordinate employees. For Andreas Sternig Supervisor=1 which is the IDof the IT-Director.
First, I want to answer the question: Who are the supervisors for each employee?
I'm going to solve this problem by recursively querying the dataset.